+++ frontend/release.config.js
@@ -1,16 +1,15 @@
-module.exports = {
+export default {
   branches: ['master'],
   tagFormat: 'frontend-v${version}',
   plugins: [
     '@semantic-release/commit-analyzer',
     '@semantic-release/release-notes-generator',
     ['@semantic-release/changelog', { changelogFile: 'CHANGELOG.md' }],
-    '@semantic-release/npm',
     [
       '@semantic-release/git',
       {
         assets: ['CHANGELOG.md', 'package.json'],
-        message: 'chore(release): ${nextRelease.version} [skip ci]',
+        message: 'chore(frontend): ${nextRelease.version} [skip ci]',
       },
     ],
     '@semantic-release/github',