+++ frontend/CHANGELOG.md
@@ -1,9 +1,8 @@
 ## [1.0.1](https://github.com/tamasjuhasz84/minicms/compare/v1.0.0...v1.0.1) (2025-07-05)
 
-
 ### Bug Fixes
 
-* prevent push conflict in frontend release ([5b42e67](https://github.com/tamasjuhasz84/minicms/commit/5b42e6776b8be0fc444b605ce362c3f471a719db))
+- prevent push conflict in frontend release ([5b42e67](https://github.com/tamasjuhasz84/minicms/commit/5b42e6776b8be0fc444b605ce362c3f471a719db))
 
 # Változásnapló